Move most files from `data/` to `utils/`
Only `site.js` is used in content files (Markdown),
so it'll remain under `data/`.

Moving the files out from `data/` simplifies the global scope.
Previously they had to be under `data/`
because JS files can't be imported from Pug files,
so I had to add the stuff to the global scope.

`utils/utils.js` looks silly, but I'll tackle it soon.
