+++ _sample-content/feeds/all/all.11tydata.js
@@ -1,5 +1,5 @@
-const entity = require('../../../data/entity')
 const site = require('../../../data/site')
+const entity = require('../../../utils/entity')
 
 module.exports = {
   collectionsKey: 'allContentPages',
