+++ components/PostMeta.js
@@ -1,7 +1,7 @@
 const { html } = require('htm/preact')
 
-const datetime = require('../data/datetime')
-const { slugify } = require('../data/slugify')
+const datetime = require('../utils/datetime')
+const { slugify } = require('../utils/slugify')
 
 module.exports = ({ date, updated, tags = [] }) => html`
   <p class="!mb-0">
