+++ utils/entity.js
